fix[DE]: Join consistency exporters
